# Individuals/Individuals/src/main/java/com/artem/individuals/dto/request/TopUpConfirmRequestDto.java
package com.artem.individuals.dto.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.util.UUID;
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class TopUpConfirmRequestDto {
    private UUID userUid;
    private UUID walletUid;
    private BigDecimal amount;
    private String comment;

}